Add IdSetChunk(Finder|Insert)
This will scan the table for all currently existing ids, and then
insert stride count of ids into the destination table on each run.

This is intended for use on tables where ids of new records may not
monotonically increase and/or where the id space is non-contiguous.

This is not appropriate to use on tables with "too many" rows, as it
will attempt to pull the entire id space in a single query.